admin/requirements.txt
Flask==0.10.1
mysql-connector
requests
